+++ SightKeeper.Domain.Model/Profile/ProfileItemClass.cs
@@ -5,9 +5,9 @@ namespace SightKeeper.Domain.Model;
 public sealed class ProfileItemClass
 {
     public ItemClass ItemClass { get; private set; }
-    public int Index { get; internal set; }
+    public byte Index { get; internal set; }
 
-    public ProfileItemClass(ItemClass itemClass, int index)
+    public ProfileItemClass(ItemClass itemClass, byte index)
     {
         ItemClass = itemClass;
         Index = index;
